move workitems to their own module
