--- drools/drools-core/src/main/java/org/drools/core/process/instance/impl/DefaultWorkItemManagerFactory.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.drools.core.process.instance.impl;
-
-import java.io.Externalizable;
-import java.io.IOException;
-import java.io.ObjectInput;
-import java.io.ObjectOutput;
-
-import org.drools.core.common.InternalKnowledgeRuntime;
-import org.drools.core.process.instance.KogitoWorkItemManager;
-import org.drools.core.process.instance.WorkItemManagerFactory;
-
-public class DefaultWorkItemManagerFactory implements Externalizable, WorkItemManagerFactory {
-
-    public KogitoWorkItemManager createWorkItemManager( InternalKnowledgeRuntime kruntime) {
-        return new KogitoDefaultWorkItemManager(kruntime);
-    }
-
-    public void readExternal(ObjectInput in) throws IOException, ClassNotFoundException {
-    }
-
-    public void writeExternal(ObjectOutput out) throws IOException {
-    }
-
-}
