+++ jbpm/jbpm-bpmn2/src/main/java/org/jbpm/bpmn2/xpath/XPATHAssignmentAction.java
@@ -23,10 +23,10 @@ import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathExpression;
 import javax.xml.xpath.XPathFactory;
 
-import org.drools.core.process.instance.KogitoWorkItem;
 import org.jbpm.process.instance.impl.AssignmentAction;
 import org.jbpm.workflow.core.node.Assignment;
 import org.kie.api.runtime.process.ProcessContext;
+import org.kie.kogito.process.workitems.KogitoWorkItem;
 import org.w3c.dom.Attr;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
