+++ jbpm/jbpm-bpmn2/src/test/resources/META-INF/kie.properties.conf
@@ -1,1 +1,2 @@
 drools.workDefinitions = BPMN2WorkDefinitions.conf WorkDefinitions.conf
+drools.workItemManagerFactory=org.jbpm.bpmn2.test.TestWorkItemManagerFactory
