+++ jbpm/jbpm-flow/src/main/java/org/jbpm/process/instance/impl/humantask/BaseHumanTaskLifeCycle.java
@@ -19,7 +19,6 @@ import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-import org.drools.core.process.instance.KogitoWorkItemManager;
 import org.jbpm.process.instance.impl.humantask.phases.Claim;
 import org.jbpm.process.instance.impl.humantask.phases.Release;
 import org.jbpm.process.instance.impl.humantask.phases.Skip;
@@ -35,6 +34,7 @@ import org.kie.kogito.process.workitem.LifeCyclePhase;
 import org.kie.kogito.process.workitem.NotAuthorizedException;
 import org.kie.kogito.process.workitem.Policy;
 import org.kie.kogito.process.workitem.Transition;
+import org.kie.kogito.process.workitems.KogitoWorkItemManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -121,7 +121,7 @@ public class BaseHumanTaskLifeCycle implements LifeCycle<Map<String, Object>> {
         if (targetPhase.isTerminating()) {
             logger.debug("Target life cycle phase '{}' is terminiating, completing work item {}", targetPhase.id(), humanTaskWorkItem.getId());
             // since target life cycle phase is terminating completing work item
-            (( KogitoWorkItemManager )manager).internalCompleteWorkItem(humanTaskWorkItem);
+            ((KogitoWorkItemManager)manager).internalCompleteWorkItem(humanTaskWorkItem);
         }
         
         return data(humanTaskWorkItem);
