+++ jbpm/jbpm-flow/src/test/java/org/jbpm/test/TestWorkItemManagerFactory.java
@@ -1,0 +1,14 @@
+package org.jbpm.test;
+
+import org.drools.core.common.InternalKnowledgeRuntime;
+import org.drools.core.process.instance.WorkItemManagerFactory;
+import org.kie.kogito.process.workitems.KogitoWorkItemManager;
+import org.kie.kogito.process.workitems.impl.KogitoDefaultWorkItemManager;
+
+public class TestWorkItemManagerFactory implements WorkItemManagerFactory {
+
+    public KogitoWorkItemManager createWorkItemManager(InternalKnowledgeRuntime kruntime) {
+        return new KogitoDefaultWorkItemManager(kruntime);
+    }
+
+}
