+++ jbpm/jbpm-flow/src/test/resources/META-INF/kie.properties.conf
@@ -1,0 +1,1 @@
+drools.workItemManagerFactory=org.jbpm.test.TestWorkItemManagerFactory
