+++ jbpm/process-workitems/src/main/java/org/kie/kogito/process/workitems/KogitoWorkItemManager.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.kogito.process.workitems;
+
+import java.util.Map;
+
+public interface KogitoWorkItemManager extends WorkItemManager {
+
+    void internalExecuteWorkItem( KogitoWorkItem workItem);
+
+    void internalAddWorkItem( KogitoWorkItem workItem);
+
+    void internalAbortWorkItem(String id);
+    
+    void internalCompleteWorkItem( KogitoWorkItem workItem);
+    
+    KogitoWorkItem getWorkItem( String id);
+
+    public void signalEvent(String type, Object event, String processInstanceId);
+
+    void retryWorkItem( String workItemID, Map<String, Object> params ) ;
+
+}
