+++ jbpm/process-workitems/src/main/java/org/kie/kogito/process/workitems/WorkItemManager.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2017 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2010 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,13 +14,8 @@
  * limitations under the License.
  */
 
-package org.jbpm.process.instance.impl;
+package org.kie.kogito.process.workitems;
 
-import org.kie.api.runtime.process.ProcessContext;
-import org.kie.kogito.process.workitems.KogitoWorkItem;
+public interface WorkItemManager extends org.drools.core.process.instance.WorkItemManager {
 
-public interface AssignmentAction {
-    
-    void execute(KogitoWorkItem workItem, ProcessContext context) throws Exception;
-    
 }
